add authorization on delete and edit sambat and comment, change button color if current user upvote/downvote to sambat

+++ resources/views/sambat/details.blade.php
@@ -10,33 +10,35 @@
                     <div class="flex items-center pb-2  ">
                         <div class="w-10 h-10">
                             <img class="object-cover w-full rounded-full mr-2"
-                                src="{{ $sambat->user->profile_photo_url }}" alt="{{ $sambat->user->name }}" />
+                                src="{{ $comment->user->profile_photo_url }}" alt="{{ $comment->user->name }}" />
                         </div>
                         <div class="ml-2">
                             <div class="font-bold text-sm text-gray-600">
-                                {{ $sambat->user->name }}
+                                {{ $comment->user->name }}
                             </div>
                             <div class="text-xs">
-                                {{ $sambat->created_at }}
+                                {{ $comment->created_at }}
                             </div>
                         </div>
                     </div>
                     <div>
-                        <x-jet-dropdown align="right" width="48">
-                            <x-slot name="trigger">
-                                <button
-                                    class="flex p-1 text-sm border-2 border-transparent rounded-full focus:outline-none focus:border-gray-300 transition">
-                                    <x-icons.more-vertical width="20" height="20" />
-                                </button>
-                            </x-slot>
+                        @if (Auth::user()->id == $comment->user->id)    
+                            <x-jet-dropdown align="right" width="48">
+                                <x-slot name="trigger">
+                                    <button
+                                        class="flex p-1 text-sm border-2 border-transparent rounded-full focus:outline-none focus:border-gray-300 transition">
+                                        <x-icons.more-vertical width="20" height="20" />
+                                    </button>
+                                </x-slot>
 
-                            <x-slot name="content">
-                                <x-jet-dropdown-link class="cursor-pointer"
-                                    wire:click="deleteComments({{ $comment->id }})">
-                                    {{ __('Delete') }}
-                                </x-jet-dropdown-link>
-                            </x-slot>
-                        </x-jet-dropdown>
+                                <x-slot name="content">
+                                    <x-jet-dropdown-link class="cursor-pointer"
+                                        wire:click="deleteComments({{ $comment->id }})">
+                                        {{ __('Delete') }}
+                                    </x-jet-dropdown-link>
+                                </x-slot>
+                            </x-jet-dropdown>
+                        @endif
                     </div>
                 </div>
                 <p class="my-2 text-sm">{{ $comment->description }}</p>
